sixth commit

+++ src/btu/Task9.java
@@ -4,12 +4,12 @@ import java.util.*;
 public class Task9 {
 
     public void Task9(){
-        int[] my_array = new int[]{15, 9, -5, 30, 64, 99, 87, 5};
+        double[] my_array = new double[]{15, 9, -5, 30, 64, 99, 87, 5};
 
-        System.out.format("\n%s\nMax: %d\nMin: %d\n", Arrays.toString(my_array), maximum(my_array), minimum(my_array));
+        System.out.format("\n%s\nMax: %.2f\nMin: %.2f\n", Arrays.toString(my_array), maximum(my_array), minimum(my_array));
     }
-    private int maximum(int[] some_array){
-        int max = 0;
+    private double maximum(double[] some_array){
+        double max = 0;
 
         for(int i = 0; i < some_array.length; i++)
             if(some_array[i] > max)
@@ -17,8 +17,8 @@ public class Task9 {
 
         return max;
     }
-    private int minimum(int[] some_array){
-        int min = 0;
+    private double minimum(double[] some_array){
+        double min = 0;
 
         for(int i = 0; i < some_array.length; i++)
             if(some_array[i] < min)
